fix: urls sections and btn "ver mas"
