+++ app/src/app/templates/page.tsx
@@ -1,11 +1,9 @@
 import TemplatesComponents from '@/documents/Docs/templates';
-// import Template from '@/documents/Templates/Template';
 
 export default function releasesDocs() {
   return (
     <>
       <TemplatesComponents />
-      {/* <Template /> */}
     </>
   );
 }
