+++ app/src/app/templates/pagina-area-dos/page.tsx
@@ -1,4 +1,4 @@
-import AreaDos from '@/documents/TemplateTramite/AreaDos';
+import AreaDos from '@/documents/Templates/AreaDos';
 
 export default function TemplatePage() {
   return (
