+++ app/src/documents/Templates/AreaDos.tsx
@@ -308,27 +308,12 @@ export default function AreaDos() {
       <div>
         <nav className="tabs-box" aria-label="Navegación por pestañas">
           <ul className="nav nav-pills nav-box tabs p-0" role="tablist">
+            <TabItem id={`panel-content-docs`} title={'Docs'} activeTab={activeTab} setActiveTab={setActiveTab} />
+            <TabItem id={`panel-content-html`} title={'HTML'} activeTab={activeTab} setActiveTab={setActiveTab} />
             <TabItem
-              // key={section.id}
-              id={`panel-content-docs`}
-              title={'Docs'}
-              icon="description"
-              activeTab={activeTab}
-              setActiveTab={setActiveTab}
-            />
-            <TabItem
-              // key={section.id}
-              id={`panel-content-html`}
-              title={'HTML'}
-              icon="description"
-              activeTab={activeTab}
-              setActiveTab={setActiveTab}
-            />
-            <TabItem
-              // key={section.id}
               id={`#`}
               title={'Demo'}
-              icon="description"
+              icon="open_in_new"
               urlDemo="https://gcba.github.io/obelisco-demo/pagina_area/area_de_gobierno/op2.html"
               activeTab={activeTab}
               setActiveTab={setActiveTab}
@@ -341,7 +326,7 @@ export default function AreaDos() {
         <div className="tab-content" style={{ paddingTop: '32px' }}>
           <TabPanel id={`panel-content-docs`} activeTab={activeTab}>
             <>
-              <DocumentationTemplate sections={DOC_SECTIONS} />
+              <DocumentationTemplate sections={DOC_SECTIONS} type="ux" />
             </>
           </TabPanel>
 
@@ -349,10 +334,6 @@ export default function AreaDos() {
             <>
               <CodeViewTemplate template={AREA_2} />
             </>
-          </TabPanel>
-
-          <TabPanel id={'panel-content-react'} activeTab={activeTab}>
-            <></>
           </TabPanel>
         </div>
       </div>
